+++ compiler-tests/src/test/data/dump/ir/dependencygraph/CreatorParamsDoNotGetChunked.kt
@@ -1,3 +1,4 @@
+// CHUNK_FIELD_INITS: true
 // https://github.com/ZacSweers/metro/issues/721
 class ProvidedExample