+++ gradle-plugin/src/main/kotlin/dev/zacsweers/metro/gradle/MetroGradleSubplugin.kt
@@ -124,6 +124,7 @@ public class MetroGradleSubplugin : KotlinCompilerPluginSupportPlugin {
         add(lazyOption("enable-scoped-inject-class-hints", extension.enableScopedInjectClassHints))
         add(lazyOption("transform-providers-to-private", extension.transformProvidersToPrivate))
         add(lazyOption("shrink-unused-bindings", extension.shrinkUnusedBindings))
+        add(lazyOption("chunk-field-inits", extension.chunkFieldInits))
         add(lazyOption("public-provider-severity", extension.publicProviderSeverity))
         add(
           lazyOption(